last fixes

+++ src/main/java/blog/api/response/StatisticResponse.java
@@ -8,10 +8,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class StatisticResponse {
-    private int postsCount;
-    private int likesCount;
-    private int dislikesCount;
-    private int viewsCount;
-    private long firstPublication;
+    private Long postsCount;
+    private Long likesCount;
+    private Long dislikesCount;
+    private Long viewsCount;
+    private Long firstPublication;
 }
